+++ sedov_blast_wave/main.cpp
@@ -1,4 +1,18 @@
-#include "sedov.hpp"
+#include "../wash_mockapi.hpp"
+#include "sedov_density_kernel.hpp"
+#include "sedov_eos_kernel.hpp"
+#include "sedov_iad_kernel.hpp"
+#include "sedov_init.hpp"
+#include "sedov_momentum_energy_kernel.hpp"
+
+void force_kernel(wash::Particle& p, std::vector<wash::Particle>& neighbours) {
+    compute_density(p, neighbours);
+    compute_eos_hydro_std(p);
+    compute_iad(p, neighbours);
+    compute_momentum_energy_std(p, neighbours);
+}
+
+void update_kernel(wash::Particle& p) {}
 
 int main(int argc, char **argv) {
     wash::set_precision("double");
