+++ sedov_blast_wave/sedov_consts.hpp
@@ -4,5 +4,6 @@
 #include <cmath>
 
 extern const double k;
+extern const size_t ngmax;
 void init_wh();
 double lookup_wh(const double v);
