+++ sedov_blast_wave/sedov_density_kernel.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../wash_mockapi.hpp"
+#include "sedov_box.hpp"
+#include "sedov_consts.hpp"
+
+void compute_density(wash::Particle& p, const std::vector<wash::Particle>& neighbours);
