+++ sedov_blast_wave/sedov_eos_kernel.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "../wash_mockapi.hpp"
+
+void compute_eos_hydro_std(wash::Particle& p);
