+++ sedov_blast_wave/sedov_init.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "../wash_mockapi.hpp"
+
+void init();
